Move UncivSlider
