+++ core/src/com/unciv/ui/screens/victoryscreen/VictoryScreenReplay.kt
@@ -7,7 +7,7 @@ import com.badlogic.gdx.utils.Timer
 import com.unciv.models.UncivSound
 import com.unciv.ui.components.Fonts
 import com.unciv.ui.components.widgets.TabbedPager
-import com.unciv.ui.components.UncivSlider
+import com.unciv.ui.components.widgets.UncivSlider
 import com.unciv.ui.components.YearTextUtil
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.components.extensions.setSize
